split tuners
